remove km page
